Update RateThisApp.java
Removed bug which shows dialog based on "How long days does it take from the app installation" if no of days are more like 30 days.

By debuging code i found that i got thresold in minus value so each time it shows a dialog which is wrong so I have taken thresold as long & also used L with 1000 to convert in actull millis.

And now it gives me correct result.

+++ library/src/com/kskkbys/rate/RateThisApp.java
@@ -110,7 +110,7 @@ public class RateThisApp {
 			if (mLaunchTimes >= LAUNCH_TIMES) {
 				return true;
 			}
-			int threshold = INSTALL_DAYS * 24 * 60 * 60 * 1000;	// msec
+			long threshold = INSTALL_DAYS * 24 * 60 * 60 * 1000L;	// msec
 			if (new Date().getTime() - mInstallDate.getTime() >= threshold) {
 				return true;
 			}
